+++ include/rain/networking/request-response/slave.hpp
@@ -1,19 +1,34 @@
 #pragma once
 
-#include "../client.hpp"
+#include "../slave.hpp"
 #include "socket.hpp"
 
 namespace Rain::Networking::RequestResponse {
 	template <typename RequestType, typename ResponseType>
-	class Client
+	class Slave
 			: virtual protected RequestResponse::Socket<RequestType, ResponseType>,
-				private Networking::Client {
+				private Networking::Slave {
 		public:
-		using RequestResponse::Socket<RequestType, ResponseType>::Socket;
+		// Needs to encompass same signature as Networking::Slave constructor, for
+		// use in onServerAccept.
+		Slave(Networking::Socket &socket,
+			std::size_t BUF_SZ = 16384,
+			const std::chrono::milliseconds &RECV_TIMEOUT_MS =
+				std::chrono::milliseconds(5000),
+			const std::chrono::milliseconds &SEND_MS_PER_KB =
+				std::chrono::milliseconds(5000),
+			const std::chrono::milliseconds &SEND_TIMEOUT_MS_LOWER =
+				std::chrono::milliseconds(5000))
+				: Networking::Socket(std::move(socket)),
+					RequestResponse::Socket<RequestType, ResponseType>(socket,
+						BUF_SZ,
+						RECV_TIMEOUT_MS,
+						SEND_MS_PER_KB,
+						SEND_TIMEOUT_MS_LOWER),
+					Networking::Slave(socket) {}
 
 		// Expose relevant functions.
 		using RequestResponse::Socket<RequestType, ResponseType>::close;
-		using RequestResponse::Socket<RequestType, ResponseType>::connect;
 		using RequestResponse::Socket<RequestType, ResponseType>::getFamily;
 		using RequestResponse::Socket<RequestType, ResponseType>::getNativeSocket;
 		using RequestResponse::Socket<RequestType, ResponseType>::getProtocol;